Resources/requirements.txt
torch
transformers
accelerate
librosa
torchao
protobuf
safetensors>=0.4.3
llama-cpp-python
